MovieStore/Tests/MovieStore.UnitTests/Application/MovieOperations/Commands/DeleteMovie/DeleteMovieCommandTests.cs


using FluentAssertions;
using MovieStore.Application.MovieOperations.Commands.DeleteMovie;
using MovieStore.DbOperations;
using MovieStore.UnitTests.TestSetup;
using System;
using Xunit;

namespace MovieStore.UnitTests.Application.MovieOperations.Commands.DeleteMovie
{
    public class DeleteMovieCommandTests : IClassFixture<CommonTestFixture>
    {
        private readonly MovieStoreDbContext _context;
        public DeleteMovieCommandTests(CommonTestFixture testFixture)
        {
            _context = testFixture.Context;
        }

        [Theory]
        [InlineData(10)]
        [InlineData(11)]
        [InlineData(12)]
        public void WhenMovieNotFound_InvalidOperationException_ShouldBeReturn(int movieId)
        {
            DeleteMovieCommand command = new DeleteMovieCommand(_context);
            command.MovieId = movieId;
            FluentActions.Invoking(() => command.Handle()).Should().Throw<InvalidOperationException>().And.Message.Should().Be("Silinecek film bulunamadı");
        }
    }
}
